Adds a new flag to 'get_report' and a new type of packet
New flag: SUPPORTED_LANGUAGES
When set in get_report, the worker answers with a list of objects.
They have the properties name and version.
They represent the list of the languages the worker is configured to support

New protocols: get_language_support and language_support.
get_language_support has a name and a command property to ask a worker to check a language
language_support answers the protocol above, with name, allow and version
More details on #1